Refactor sign transaction rpc method.
