+++ snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/fioprotocol/fio-wallet-snap.git"
   },
   "source": {
-    "shasum": "9iWhyjnmHiCoRdas3lVInajq/z/r/3VBTbgqzY++Hcw=",
+    "shasum": "lQ/XeLIVArYr5iivDqmlnL7EBik6Hvc2szYpjy/cg9o=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
